chore: todo-list를 보여주는 페이지를 메인으로 이동
