--- src/app/boards/layout.tsx
@@ -1,12 +1,0 @@
-import { TodoGnb } from "@/components/todo-gnb";
-
-function TodoListLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <div>
-      <TodoGnb />
-      {children}
-    </div>
-  );
-}
-
-export default TodoListLayout;
